requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
yfinance>=0.1.70
scikit-learn>=1.0.0
pytest>=6.2.5
finnhub-python>=2.4.14
alpaca-trade-api>=2.3.0
alpha_vantage>=2.3.1
